Tight integration of Channels into the library.
In the old implementation Transports were free to create their own packet headers. This created a dynamic environment but also created chaos for different transport implementations. To overcome this issue now the entire library's packet headers are based on a struct called PacketHeader. Transports are forced to use them. They can still add on top of it as they wish but the base structure of the header is nased on the struct I have created.

Within this new packet header we are currently using 2 bytes. First it channel and second is packet id. Previously writing channels into the packet were optional for the Transports. We forced it's usage. To keep the header size same as before, reather using unsigned integer for packetId I decided to use a byte. With 255 channels and 255 different packet ids, amount of packet ids you can create is same as before(if you were not using channels) which is 65k different packet ids/message handlers.